DB transaction context

+++ src/family_apiary/products/infrastructure/database/migrations/2025_06_02_01_56_34_create_purchase_requests_tables.py
@@ -55,7 +55,7 @@ def upgrade() -> None:
         sa.Column('name', sa.String(), nullable=False),
         sa.Column('description', sa.String(), nullable=False),
         sa.Column('category', sa.String(), nullable=False),
-        sa.Column('price', sa.DECIMAL(), nullable=False),
+        sa.Column('price', sa.Float, nullable=False),
         sa.Column('count', sa.INTEGER(), nullable=False),
         sa.ForeignKeyConstraint(
             ['purchase_request_id'],
